Move gateway trace link to SpanModelCostBadge and extract linkedGatewayTraceId before mlflow.* filter
The previous approach placed the link in the attributes tab, but mlflow.*
attributes are filtered out in normalizeSpan before reaching UI components.
This moves the gateway trace link to display alongside model/cost badges
by extracting linkedGatewayTraceId as a first-class span node field.
