+++ mlflow/server/js/src/shared/web-shared/model-trace-explorer/ModelTrace.types.ts
@@ -299,6 +299,7 @@ export interface ModelTraceSpanNode extends TimelineTreeNode, Pick<ModelTraceSpa
   traceId: string;
   modelName?: string;
   cost?: SpanCostInfo;
+  linkedGatewayTraceId?: string;
 }
 
 export type ModelTraceExplorerTab = 'chat' | 'content' | 'attributes' | 'events';
